Move to normal malloc things
"by default" LVGL has it's own memory management. This change moves us more to the malloc/free calls that linux os has.
This works better for us, as there seems to be an issue on mac/linux with the lvgl memory management.

This will probably not work on ESP or other embedded devices, which makes it more logical why lvgl choose this as the default.