Introduce UIComponent
This should be the base class for any LVGL related stuff we do. We can store useful methods like `createLabel()` and can consider it for styling stuff later on.

We still have UIEntity, which is the uicomponent that is linked to the HAEntity. We will later change that into 'multiple haentities' as that sometimes is needed too.

This class model gives us a bit more flexibility to have useful LVGL stuff in a central place, while not being tied to 'always have HAentity'.