run the openwrt-armel build in a debian:12 docker container
I previously noticed (but never fully investigated) annoying differences
between the GH ubuntu runtime environment and the nektos/act one.
Then GH did an update to the ubuntu runner and the build broke there too.

Switching to the debian:12 docker image hopefully gives us a more stable environment.

However, the actual stability change in this commit appears to be the removal of the second package dir inside the src tree.
This commit still switches us to Debian 12 for potential stability in general.